+++ src/main/java/org/musicplace/global/security/controller/AuthController.java
@@ -4,8 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.musicplace.global.security.config.CustomUserDetails;
 import org.musicplace.global.security.dto.LoginResponseDto;
 import org.musicplace.global.security.jwt.JwtTokenUtil;
-import org.musicplace.member.domain.SignInEntity;
-import org.musicplace.member.dto.LoginRequestDto;
+import org.musicplace.global.security.dto.LoginRequestDto;
 import org.musicplace.member.service.SignInService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
