+++ src/main/java/org/musicplace/global/security/dto/LoginRequestDto.java
@@ -1,11 +1,19 @@
 package org.musicplace.global.security.dto;
 
+import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
+import lombok.NoArgsConstructor;
 
 @Getter
-@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class LoginRequestDto {
-    private String username;
-    private String password;
+    private String member_id;
+    private String pw;
+
+    @Builder
+    public LoginRequestDto(String member_id, String pw) {
+        this.member_id = member_id;
+        this.pw = pw;
+    }
 }
