--- src/main/java/org/musicplace/member/dto/LoginRequestDto.java
@@ -1,19 +1,0 @@
-package org.musicplace.member.dto;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LoginRequestDto {
-    private String member_id;
-    private String pw;
-
-    @Builder
-    public LoginRequestDto(String member_id, String pw) {
-        this.member_id = member_id;
-        this.pw = pw;
-    }
-}
