+++ src/main/resources/application-loadtest.yml
@@ -18,12 +18,16 @@ spring:
   jpa:
     open-in-view: true
     hibernate:
-      ddl-auto: validate              # 테스트 시 스키마 변경 방지
+      ddl-auto: create-drop
     show-sql: false                   # 부하 테스트 시 로그 최소화
     properties:
       hibernate.format_sql: false
       hibernate.dialect: org.hibernate.dialect.MySQLDialect
       hibernate.jdbc.batch_size: 20   # 배치 처리 최적화
+
+  sql:
+    init:
+      mode: always
 
   data:
     redis:
