+++ src/main/resources/application.yml
@@ -14,6 +14,10 @@ spring:
       hibernate.format_sql: true
       dialect: org.hibernate.dialect.MySQLDialect
 
+  sql:
+    init:
+      mode: always
+
   data:
     redis:
       host: ${REDIS_HOST}
